Move models from directory db to models
