+++ src/db/__init__.py
@@ -1,8 +1,8 @@
 from .database import Base, engine, database
-from .user import User
-from .user_room import UserRoom
-from .room_members import RoomMembers
-from .message import Message
+from .models.user import User
+from .models.user_room import UserRoom
+from .models.room_members import RoomMembers
+from .models.message import Message
 
 # Создаем таблицы
 Base.metadata.create_all(bind=engine)
