+++ src/db/models/room_members.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Column, ForeignKey, Integer
 from sqlalchemy.orm import relationship
 
-from .database import Base
+from ..database import Base
 
 
 class RoomMembers(Base):
